Add CSS-Zen-Garden.css to .gitignore
This is a stylesheet for the homepage of csszengarden.com (a page
recommended for reference for this project). When I inspected the
stylesheet it was all on a single line (despite carriage returns),
making it impossible to examine, so I imported the text into a file.
This sheet is not meant to be used for any project as it is the sole
property of the original author. It is here for reference only so it
can inform my learning.